BBDD

+++ database/migrations/2018_06_28_213634_create_almacenes_table.php
@@ -15,7 +15,7 @@ class CreateAlmacenesTable extends Migration
     {
         Schema::create('almacenes', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('nivel');
+            $table->unsignedTinyInteger('nivel');
             $table->unsignedBigInteger('capacidad');
         });
     }
