+++ database/migrations/2018_06_29_220838_create_construcciones_table.php
@@ -17,7 +17,7 @@ class CreateConstruccionesTable extends Migration
             $table->increments('id');
             $table->string('codigo')->index();
             $table->string('categoria');
-            $table->integer('nivel')->default(0);
+            $table->unsignedTinyInteger('nivel')->default(0);
         });
     }
