+++ database/migrations/2018_07_14_124453_create_en_investigaciones_table.php
@@ -15,8 +15,8 @@ class CreateEnInvestigacionesTable extends Migration
     {
         Schema::create('en_investigaciones', function (Blueprint $table) {
             $table->increments('id');
-            $table->decimal('personal', 14, 2);
-            $table->integer('nivel');
+            $table->decimal('personal', 13, 2);
+            $table->unsignedTinyInteger('nivel');
             $table->string('accion');
             $table->timestamps();
             $table->timestamp('finished_at')->nullable();
