+++ database/migrations/2018_08_27_204914_create_cualidades_disenios_table.php
@@ -16,7 +16,7 @@ class CreateCualidadesDiseniosTable extends Migration
         Schema::create('cualidades_disenios', function (Blueprint $table) {
             $table->increments('id');
             $table->string('codigo')->index();
-            $table->integer('cantidad');
+            $table->unsignedSmallInteger('cantidad');
             $table->string('categoria');
         });
     }
